+++ flecs_ecs/Cargo.toml
@@ -12,6 +12,7 @@ categories = ["game-development", "api-bindings"]
 [dependencies]
 #debug-here = "0.2.2" #temp for development
 flecs_ecs_derive = { path = "../flecs_ecs_derive" }
+flecs_ecs_sys = { path = "../flecs_ecs_sys" }
 seq-macro = "0.3.5" #temp for testing / development
 rand = "0.8.5"  #temp for testing / development
 libc = "0.2.153"
@@ -38,7 +39,6 @@ regex = "1.9.5"
 #https://github.com/rust-lang/rust-bindgen/issues/1966
 
 [features]
-flecs_generate_bindings = [] 
 flecs_app = []
 flecs_pipeline = []
 flecs_system = []
@@ -69,8 +69,8 @@ default = [
     "flecs_rules",
     "flecs_app",
     "flecs_pipeline",
-    "flecs_generate_bindings"
 ]
 
 [target.'cfg(any())'.dependencies]
 flecs_ecs_derive = { version = "=0.1.0", path = "../flecs_ecs_derive" }
+flecs_ecs_sys = { version = "=0.1.0", path = "../flecs_ecs_sys" }
